+++ CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to OverType will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.1.7] - 2025-01-20
+
+### Security
+- Fixed XSS vulnerability where javascript: protocol links could execute arbitrary code (#25)
+  - Added URL sanitization to block dangerous protocols (javascript:, data:, vbscript:, etc.)
+  - Safe protocols allowed: http://, https://, mailto:, ftp://, ftps://
+  - Relative URLs and hash links continue to work normally
+  - Dangerous URLs are neutralized to "#" preventing code execution
+
 ## [1.1.6] - 2025-01-20
 
 ### Fixed